Agregar validaciones en los métodos eliminarCliente y eliminarProducto para verificar la existencia de los registros antes de eliminarlos, lanzando una excepción si no se encuentran.

+++ src/main/java/com/rubenrf/cafeteria_app/service/impl/ClienteServiceImpl.java
@@ -48,6 +48,9 @@ public class ClienteServiceImpl implements ClienteService {
 
     @Override
     public void eliminarCliente(Long idCliente) {
+        if (!clienteRepository.existsById(idCliente)) {
+            throw new EntityNotFoundException("Cliente #" + idCliente + " no encontrado.");
+        }
         clienteRepository.deleteById(idCliente);
     }
